[FIX] form 태그로 변경
- 이중 제출되는 거 방지하기 위해 변경
- 파일 분리
- 대댓글 작성 api 없는 거라 제거
